fix: type total charge as int so external calculators accept it (#161)
* fix: type total charge as int so external calculators accept it

A system's total charge is physically an integer, but the benchmark input
models (`Molecule`, `Fragment`, `Conformer`, `Reaction`, `TautomerPair`,
`MolecularSystem`, …) and the `STRUCTURE_CHARGES` tables typed/stored it as a
float, and the MD/analysis paths assigned `atoms.info["charge"] = float(...)`.

Native mlip force fields tolerate a float charge, but external ASE calculators
(notably fairchem/UMA's `FAIRChemCalculator`, omol task) reject it with
`Invalid type for charge: <class 'float'>. Charge must be an integer ...`,
which made every molecular benchmark (folding_stability, sampling, etc.) error
out for UMA. This mirrors NEB, which already types `charge: int`.

Fix: type the charge fields and `STRUCTURE_CHARGES` as `int` (pydantic coerces
the integral source values at parse time) and drop the `float(...)` wrappers at
the `atoms.info["charge"] = ...` assignment sites. No behavioural change for
native models; unblocks external calculators.


* chore: trigger claude review

* docs: update changelog

* feat: bump version

---------

+++ src/mlipaudit/benchmarks/bond_length_distribution/bond_length_distribution.py
@@ -69,7 +69,7 @@ class Molecule(BaseModel):
     coordinates: list[tuple[float, float, float]]
     pattern_atom_indices: tuple[int, int]
     reference_bond_distance: float
-    charge: float
+    charge: int
     smiles: str
 
 
@@ -196,7 +196,7 @@ class BondLengthDistributionBenchmark(Benchmark):
                 symbols=molecule.atom_symbols,
                 positions=molecule.coordinates,
             )
-            atoms.info["charge"] = float(molecule.charge)
+            atoms.info["charge"] = molecule.charge
             atoms.info["spin"] = DEFAULT_SPIN
             simulation_state = run_simulation(atoms, self.force_field, **md_kwargs)
