+++ src/mlipaudit/benchmarks/stability/stability.py
@@ -122,15 +122,15 @@ BOX_SIZES = {
 # Total charge per structure. All systems are treated as closed-shell singlets
 # (spin multiplicity = 1). Values are estimates from system composition and
 # parity-checked against the observed electron count.
-STRUCTURE_CHARGES: dict[str, float] = {
-    "Small_molecule_HCNO": 0.0,
-    "Small_molecule_Sulfur": 0.0,
-    "Small_molecule_Halogen": 0.0,
-    "Peptide_HCNO": 1.0,
-    "Peptide_cys": 0.0,
-    "Protein": 7.0,
-    "Peptide_solvated": 0.0,
-    "Peptide_solvated_ions": 0.0,
+STRUCTURE_CHARGES: dict[str, int] = {
+    "Small_molecule_HCNO": 0,
+    "Small_molecule_Sulfur": 0,
+    "Small_molecule_Halogen": 0,
+    "Peptide_HCNO": 1,
+    "Peptide_cys": 0,
+    "Protein": 7,
+    "Peptide_solvated": 0,
+    "Peptide_solvated_ions": 0,
 }
 
 STRUCTURE_NAMES = list(STRUCTURES.keys())
@@ -435,7 +435,7 @@ class StabilityBenchmark(Benchmark):
             logger.info("Running MD for %s", structure_name)
             xyz_filename = STRUCTURES[structure_name]["xyz"]
             atoms = ase_read(self.data_input_dir / self.name / xyz_filename)
-            atoms.info["charge"] = float(STRUCTURE_CHARGES[structure_name])
+            atoms.info["charge"] = STRUCTURE_CHARGES[structure_name]
             atoms.info["spin"] = DEFAULT_SPIN
 
             if structure_name in BOX_SIZES:
